requirements.txt
pyglet
requests
python-dotenv
